requirements.txt
anyio==4.7.0
autopep8==2.3.1
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
python-dotenv==1.0.1
python-telegram-bot==21.9
setuptools==75.6.0
sniffio==1.3.1
typing_extensions==4.12.2
wheel==0.45.1
yt-dlp==2024.12.23
